Closes GH-0026 review change - bumping up normalizeAppPath(conf) to JobUtils class
