+++ core/src/main/java/org/apache/oozie/servlet/BaseJobServlet.java
@@ -37,6 +37,7 @@ import org.apache.oozie.service.HadoopAccessorException;
 import org.apache.oozie.service.HadoopAccessorService;
 import org.apache.oozie.service.Services;
 import org.apache.oozie.service.XLogService;
+import org.apache.oozie.util.JobUtils;
 import org.apache.oozie.util.XConfiguration;
 import org.apache.oozie.util.XLog;
 import org.json.simple.JSONObject;
@@ -107,7 +108,7 @@ public abstract class BaseJobServlet extends JsonRestServlet {
             Configuration conf = new XConfiguration(request.getInputStream());
             stopCron();
             checkAuthorizationForApp(getUser(request), conf);
-            normalizeAppPath(conf);
+            JobUtils.normalizeAppPath(conf);
             reRunJob(request, response, conf);
             startCron();
             response.setStatus(HttpServletResponse.SC_OK);
@@ -176,54 +177,6 @@ public abstract class BaseJobServlet extends JsonRestServlet {
         catch (AuthorizationException ex) {
             XLog.getLog(BaseJobServlet.class).info("AuthorizationException ", ex);
             throw new XServletException(HttpServletResponse.SC_UNAUTHORIZED, ex);
-        }
-    }
-
-    /**
-     * Normalize appPath in conf - If it's not jobs via http submission, after normalization appPath always
-     * points to xml definition file. <p/>
-     *
-     * @param conf configuration.
-     * @throws IOException thrown if normalization can not be done properly.
-     */
-    public static void normalizeAppPath(Configuration conf) throws IOException {
-        if (conf.get(XOozieClient.IS_PROXY_SUBMISSION) != null) { // do nothing for proxy submission job;
-            return;
-        }
-
-        String wfPathStr = conf.get(OozieClient.APP_PATH);
-        String coordPathStr = conf.get(OozieClient.COORDINATOR_APP_PATH);
-        String appPathStr = wfPathStr != null ? wfPathStr : coordPathStr;
-
-        String user = conf.get(OozieClient.USER_NAME);
-        String group = conf.get(OozieClient.GROUP_NAME);
-        FileSystem fs = null;
-        try {
-            fs = Services.get().get(HadoopAccessorService.class).createFileSystem(user, group, new Path(appPathStr).toUri(), conf);
-        }
-        catch (HadoopAccessorException ex) {
-            throw new IOException(ex.getMessage());
-        }
-
-        Path appPath = new Path(appPathStr);
-        String normalizedAppPathStr = appPathStr;
-        if (!fs.exists(appPath)) {
-            throw new IOException("Error: " + appPathStr + " does not exist");
-        }
-
-        FileStatus fileStatus = fs.getFileStatus(appPath);
-        Path appXml = appPath;
-        // Normalize appPath here - it will always point to a workflow/coordinator xml definition file;
-        if (fileStatus.isDir()) {
-            appXml = new Path(appPath, (wfPathStr != null)? "workflow.xml" : "coordinator.xml");
-            normalizedAppPathStr = appXml.toString();
-        }
-
-        if (wfPathStr != null) {
-            conf.set(OozieClient.APP_PATH, normalizedAppPathStr);
-        }
-        else if (coordPathStr != null) {
-            conf.set(OozieClient.COORDINATOR_APP_PATH, normalizedAppPathStr);
         }
     }
