+++ core/src/main/java/org/apache/oozie/servlet/BaseJobsServlet.java
@@ -27,6 +27,7 @@ import org.apache.oozie.client.OozieClient;
 import org.apache.oozie.client.rest.RestConstants;
 import org.apache.oozie.service.Services;
 import org.apache.oozie.service.WorkflowAppService;
+import org.apache.oozie.util.JobUtils;
 import org.apache.oozie.util.XConfiguration;
 import org.json.simple.JSONObject;
 
@@ -86,7 +87,7 @@ public abstract class BaseJobsServlet extends JsonRestServlet {
 
         validateJobConfiguration(conf);
         BaseJobServlet.checkAuthorizationForApp(getUser(request), conf);
-        BaseJobServlet.normalizeAppPath(conf);
+        JobUtils.normalizeAppPath(conf);
 
         JSONObject json = submitJob(request, conf);
         startCron();
